small change to OS X sdl_mixer handling, updated os x version numbers
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@3932 01821946-851b-0410-b7f4-f1477c70e57a

+++ main/gameargs.cpp
@@ -258,7 +258,7 @@ if ((t = FindArg ("-sdl_mixer")))
 	gameOptions [0].sound.bUseSDLMixer = NumArg (t, 1);
 #endif
 #ifdef __macosx__
-	unsigned int address = (unsigned int)(Mix_OpenAudio); 
+	uintptr_t address = (uintptr_t)(Mix_OpenAudio); 
 	if(address == 0u) { 
 		// the SDL_mixer framework is not present,
 		// so regardless of what conf.h or d2x.ini says,
